h264 works

+++ init-scripts/sqript.sql
@@ -1,6 +1,6 @@
 CREATE TABLE videos (
     id SERIAL PRIMARY KEY,
-    video_size INTEGER,
-    encoding_time INTEGER,
-    decoding_time INTEGER
+    video_size FLOAT,
+    encoding_time FLOAT,
+    decoding_time FLOAT
 );
